requirements.txt
django
djangorestframework
drf-yasg==1.21.7